Move providers in specific folder
